add support to decode and encode the orientation parameter
The documentation in the firmware mentioned that the orientation
string should be a three letter code with uppercase for "positive axis"
and lowercase for "negative axis". However, the configuration read
was reporting a string of the form +X+Y+Z

I went on to implement a proper decoding/encoding scheme to finally
realize that the documentation was wrong, and was expecting us to
set the orientation with the same +X+Y+Z

I elected to keep the properly decoded value, though, as it feels
cleaner than moving a string along.