feat: implement imu_aceinna_openimu_nmea_publisher
It acts as a proxy between the IMU and another client (i.e. Syskit),
publishing state as NMEA to a configurable target. This will allow us
to send NMEA state to devices that require it (e.g. the satcoms) without
the involvement from the Syskit system - making sure that that part
works even if the system is down.